+++ samtranslator/internal/data/managed_policies.json
